+++ public/src/app/tt-core/components/app.component.ts
