+++ public/src/app/tt-login/services/login.service.ts
@@ -3,7 +3,7 @@ import { Headers, Http }       from '@angular/http';
 
 import 'rxjs/add/operator/toPromise';
 
-import { User }  from './user';
+import { User }  from './../../models/user';
 
 @Injectable()
 export class LoginService {
